# postalpool: better pooling by default, offer configurable knobs