```
goos: darwin
goarch: amd64
pkg: github.com/moov-io/watchman/internal/postalpool
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz

               │ ./internal/postalpool/before.txt │ ./internal/postalpool/after.txt │
               │              sec/op              │    sec/op     vs base           │
_PostalPool-16                       76.08µ ± ∞ ¹   55.66µ ± ∞ ¹  ~ (p=1.000 n=1) ²
Improvement: ~26.8% faster

               │ ./internal/postalpool/before.txt │ ./internal/postalpool/after.txt  │
               │               B/op               │     B/op       vs base           │
_PostalPool-16                      18.75Ki ± ∞ ¹   14.46Ki ± ∞ ¹  ~ (p=1.000 n=1) ²
Improvement: ~22.9% less memory

               │ ./internal/postalpool/before.txt │ ./internal/postalpool/after.txt │
               │            allocs/op             │  allocs/op    vs base           │
_PostalPool-16                        152.0 ± ∞ ¹    143.0 ± ∞ ¹  ~ (p=1.000 n=1) ²
Improvement: ~5.9% fewer allocations
```